Move firebase-specific matters to firebase folder
