+++ src/firebase/api.js
@@ -1,0 +1,3 @@
+import { functions } from './index';
+
+export const createUser = functions.httpsCallable('createUser');
